closes https://github.com/helm/helm/issues/13241

**What this PR does / why we need it**:
Added `--skip-download-if-exists` into `dependency build/update` commands, it allows you to specify whether charts should be downloaded if they already exist locally.  
I'll add tests if you're ok with the changes

**If applicable**:
- [Y] this PR contains user facing changes (the `docs needed` label should be applied if so)
- [N] this PR contains unit tests
- [N] this PR has been tested for backwards compatibility