@markmcmurray-sysdig 2024-08-12
Thanks for the effort here though - it's on the right track to make a performance improvement in Helm imo