@benoittgt 2025-11-03
Hello @idsulik 

LGTM. I did a small benchmark and it is great.

```
Benchmark 1: WITHOUT --skip-download-if-exists
---------------------------------------
Run 1/5: 1800ms
Run 2/5: 1808ms
Run 3/5: 1791ms
Run 4/5: 1789ms
Run 5/5: 1841ms
Average: 1805ms

Benchmark 2: WITH --skip-download-if-exists
---------------------------------------
Run 1/5: 50ms
Run 2/5: 69ms
Run 3/5: 47ms
Run 4/5: 52ms
Run 5/5: 50ms
Average: 53ms

=========================================
RESULTS SUMMARY
=========================================
Without flag: 1805ms (average)
With flag:    53ms (average)
Improvement:  1752ms (97% faster)
```