@idsulik 2025-12-07
@markmcmurray-sysdig hi! Could you please check the R?